exchange_scrape/requirements.txt
redis == 3.5.3
websockets == 8.1
python-dotenv == 0.15.0
